% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesSLAM-package.R
\docType{package}
\name{BayesSLAM}
\alias{BayesSLAM}
\alias{BayesSLAM-package}
\title{BayesSLAM: Bayesian Statistical Learning Approach for Microbiome}
\description{
The BayesSLAM package provides a collection of microbiome data analysis functions.
}
\section{Analysis functions}{

The mbTest functions is the entry point.
}

\section{Visualization functions}{

The drawPhyloTree functions can plot cladogram.
}

\author{
Xiaowei Zhan <Xiaowei.Zhan@swmed.edu>
}
