% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_other_method.R
\name{checkPrereq}
\alias{checkPrereq}
\title{Check if the required packages are installed}
\usage{
checkPrereq(testName)
}
\arguments{
\item{testName}{a name of a statistical test}
}
\value{
TRUE only all required packages are installed
}
\description{
Check if the required packages are installed
}
\examples{
checkPrereq("deseq")
checkPrereq("mispu")
}
