% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function-computeMetaphlanTree.R
\name{computeMetaphlanTree}
\alias{computeMetaphlanTree}
\title{Compute a subset of tree for a count table}
\usage{
computeMetaphlanTree(taxa, tree = NULL)
}
\arguments{
\item{taxa}{character vector (e.g. rownames of otu_table, provided by MetaPhlAn)}

\item{tree}{R phylo object}
}
\value{
a phylo tree
}
\description{
Based on a MetaPhlAn tree, only keep the tip nodes that appear in the count table row names.
}
\examples{
library(ape)
prefix = system.file("extdata", package = "BayesSLAM")
load(file.path(prefix, "Castro-NallarE_2015.Rdata"))
countsTable <- otuTable
tree <- computeMetaphlanTree(countsTable)
}
