% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function-computeUnifrac.R
\name{computeUnifrac}
\alias{computeUnifrac}
\title{Compute Unifrac distances using GUnifrac for MetaPhlAn count table}
\usage{
computeUnifrac(otu.tab.rff, tree = NULL, alpha = c(0, 0.5, 1))
}
\arguments{
\item{otu.tab.rff}{a matrix, rownames are metaphlan tip species}
}
\value{
a list of UniFrac distance matrix using alpha 0, 0.5 and 1
}
\description{
The tree structure are provided by curatedMetagenomicData package
}
\examples{
library(GUniFrac)
prefix = system.file("extdata", package = "BayesSLAM")
load(file.path(prefix, "Castro-NallarE_2015.Rdata"))
countsTable <- otuTable
# otu.tab.rff <- Rarefy(countsTable)$otu.tab.rff
otu.tab.rff <- countsTable
unifracs = computeUnifrac(otu.tab.rff)
   
}
