% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dm_model_estimator}
\alias{dm_model_estimator}
\title{DM model}
\usage{
dm_model_estimator(Y, z, iter, S, aggregate, store)
}
\arguments{
\item{Y}{count matrix (taxa by sample)}

\item{z}{phenotype vector}

\item{iter}{number of iterations}

\item{S}{a matrix used to aggregate upper taxonomy classes}

\item{aggregate}{whether to perform analysis on upper taxonomy classes}

\item{store}{whether to store additional Bayesian statistics in addition to PPI (posterior probability of inclusion)}
}
\description{
DM model
}
