\name{mbTest-package}
\alias{mbTest-package}
\alias{mbTest}
\docType{package}
\title{
\packageTitle{mbTest}
}
\description{
\packageDescription{mbTest}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{mbTest}
\packageIndices{mbTest}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{mbTest}

Maintainer: \packageMaintainer{mbTest}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
