% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_other_method.R
\name{mbTest}
\alias{mbTest}
\title{Unified statistical tests for microbiome data}
\usage{
mbTest(testName = c("t", "anova", "wilcox", "kruskal", "deseq", "deseq2",
  "edger", "metagenomeSeq", "metagenomeSeq.zig", "permanova", "permanovaG",
  "mircat", "qcat", "zidgm", "mispu"), countsTable, pheno, cov = NULL,
  taxaTable = NULL)
}
\arguments{
\item{testName}{specify the name of the statistical test}

\item{countsTable}{OTU feature by sample matrix}

\item{pheno}{vector, factor, 2 or 3 groups}

\item{cov}{matrix, can be NULL}

\item{taxaTable}{phyloseq taxonomyTable (created by tax_table)}
}
\value{
a list of pvalue, detail and timing. The detail element stores original results
}
\description{
Unified statistical tests for microbiome data
}
\examples{
 ## two groups
if (FALSE) {
  load("~/Dropbox/xiaowei/data/real_data/curatedMetagenomicData.Rdata", verbose = TRUE)
  otuTable <- otu_list[[1]]
  taxaTable <- tax_list[[1]]
  pheno <- z_list[[1]]
} else {
  prefix = system.file("extdata", package = "BayesSLAM")
  load(file.path(prefix, "Castro-NallarE_2015.Rdata"))
}
countsTable <- otuTable
res <- mbTest("permanova", otuTable, pheno, NULL, taxaTable)
res <- mbTest("permanovaG", otuTable, pheno, NULL, taxaTable)
res <- mbTest("mircat", otuTable, pheno, NULL, taxaTable)
res <- mbTest("qcat", otuTable, pheno, NULL, taxaTable)
res <- mbTest("mispu", otuTable, pheno, NULL, taxaTable)

## three groups
if (FALSE) {
  load("~/Dropbox/xiaowei/data/real_data/curatedMetagenomicData.Rdata", verbose = TRUE)
  otuTable <- otu_list[[15]]
  taxaTable <- tax_list[[15]]
  pheno <- z_list[[15]]
} else {
  prefix = system.file("extdata", package = "BayesSLAM")
  load(file.path(prefix, "ZellerG_2014.Rdata"))
}
countsTable <- otuTable
res <- mbTest("anova", otuTable, pheno, NULL, NULL)
res <- mbTest("kruskal", otuTable, pheno, NULL, NULL)
res <- mbTest("deseq2", otuTable, pheno, NULL, NULL)
res <- mbTest("metagenomeSeq", otuTable, pheno, NULL, NULL)
res <- mbTest("metagenomeSeq.zig", otuTable, pheno, NULL, NULL)
}
