% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function-rarefy.R
\name{rarefy}
\alias{rarefy}
\title{Efficient rarefy count matrix}
\usage{
rarefy(otu)
}
\arguments{
\item{otu}{a taxa-by-sample count matrix}
}
\value{
rarefy results
}
\description{
Efficient rarefy count matrix
}
