% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{zinb_model_estimator}
\alias{zinb_model_estimator}
\title{ZINB model}
\usage{
zinb_model_estimator(Y, z, s, iter, DPP, S, aggregate, store, b, h)
}
\arguments{
\item{Y}{count matrix (taxa by sample)}

\item{z}{phenotype vector}

\item{iter}{number of iterations}

\item{S}{a matrix used to aggregate upper taxonomy classes}

\item{aggregate}{whether to perform analysis on upper taxonomy classes}

\item{store}{whether to store additional Bayesian statistics in addition to PPI (posterior probability of inclusion)}

\item{b}{a hyper parameter}
}
\description{
ZINB model
}
