% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function-drawPhyloTree.R
\name{drawPhyloTree}
\alias{drawPhyloTree}
\title{Draw phylogenetic tree in circular layout}
\usage{
drawPhyloTree(taxa, highlight = NULL, title = NULL)
}
\arguments{
\item{taxa}{phyloseq taxonomyTable class or a data frame}

\item{highlight}{a vector of node names to be colored, e.g. "k__Bacteria", 
or a data frame with optionally columns (col, size, alpha)}

\item{title}{Title for the plot}
}
\value{
an ggplot2 object
}
\description{
Draw phylogenetic tree in circular layout
}
\examples{
if (FALSE) {
  res <- curatedMetagenomicData("Castro-*metaphlan*", dryrun=FALSE) #one dataset
  taxa<-tax_table(ExpressionSet2phyloseq(res[[1]]))
} else {
  prefix = system.file("extdata", package = "BayesSLAM")
  load(file.path(prefix, "Castro-NallarE_2015.Rdata"))
  taxa <- taxaTable
}
drawPhyloTree(taxa)
drawPhyloTree(taxa, highlight = "p__Bacteroidetes")
drawPhyloTree(taxa, highlight = data.frame(taxa = c("p__Bacteroidetes", "s__Prevotella_salivae"), size = c(8, 5), col = c("red", "blue"), alpha = c(.6, .7)))
}
