% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_other_method.R
\name{mbTest}
\alias{mbTest}
\title{Deprecated - Unified statistical tests for microbiome data}
\usage{
mbTest(testName, countsTable, pheno, cov = NULL, taxaTable = NULL)
}
\arguments{
\item{testName}{specify the name of a statistical test}

\item{countsTable}{OTU feature by sample matrix}

\item{pheno}{vector, factor, 2 or 3 groups}

\item{cov}{matrix, can be NULL}

\item{taxaTable}{phyloseq taxonomyTable (created by tax_table)}
}
\value{
a list of pvalue, detail and timing. The detail element stores original results
}
